DB定数化、inquirerのmain側への移動、tmpのパッケージDL
