+++ 04.class/main.js
@@ -2,11 +2,38 @@ import Memo from "./memo.js";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 import readline from "readline";
+import inquirer from "inquirer";
 
 const memo = new Memo();
 await Memo.createTable();
 const argv = yargs(hideBin(process.argv)).argv;
 const memos = await memo.fetchMemos();
+
+const chooseMemo = (memos, messageType) => {
+  if (memos.length == 0) {
+    throw new Error("メモがありません。");
+  }
+  const message = {
+    read: "Choose a note you want to see",
+    edit: "Choose a note you want to edit",
+    delete: "Choose a note you want to delete",
+  };
+
+  const choices = memos.map((memo) => ({
+    name: memo.title,
+    value: memo,
+  }));
+
+  const prompt = inquirer.prompt([
+    {
+      type: "list",
+      name: "selectedMemo",
+      message: message[messageType],
+      choices: choices,
+    },
+  ]);
+  return prompt.then((answers) => answers.selectedMemo);
+}
 
 try {
   if (argv.l) {
@@ -14,13 +41,13 @@ try {
       console.log(memo.title);
     });
   } else if (argv.r) {
-    const chosenMemo = await memo.chooseMemo(memos, "read");
+    const chosenMemo = await chooseMemo(memos, "read");
     console.log(chosenMemo.content);
   } else if (argv.d) {
-    const chosenMemo = await memo.chooseMemo(memos, "delete");
+    const chosenMemo = await chooseMemo(memos, "delete");
     await memo.deleteMemo(chosenMemo.title);
   } else if (argv.e) {
-    const chosenMemo = await memo.chooseMemo(memos, "edit");
+    const chosenMemo = await chooseMemo(memos, "edit");
     await memo.outputMemo(chosenMemo.content);
     const editedContent = await memo.editMemo();
     await memo.updateMemo(
