+++ 04.class/package.json
@@ -11,6 +11,7 @@
   "type": "module",
   "dependencies": {
     "inquirer": "^9.2.7",
+    "tmp": "^0.2.1",
     "yargs": "^17.7.1"
   }
 }
